Fixes error in success handler for gdpr-privacy-preferences-frm submission
* Removes unused local variable
* Found `response.status` is undefined, so use `xhr.status` check instead - however may not be needed
* Force `window.location.reload` to load page from server, rather than local cache

It may be that the `200 === xhr.status` check is not required. This is the success handler function, so a successful outcome is assumed(?)

I haven't updated the `else` / `displayNotification` function, as I'm not sure what the expected contents of the `response.data...` object would be. Perhaps hardcoded with translated error strings?